Extract `Rekkyo::Type::Member` class into another file
